pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-tools-collection"
version = "1.0.0"
description = "A collection of Model Context Protocol (MCP) tools"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "MCP Tools Collection contributors", email = "maintainer@example.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

[project.optional-dependencies]
icogen = [
    "icogen-mcp>=0.1.0",
]
runcmd = [
    "runcmd-mcp>=0.1.0",
]
all = [
    "icogen-mcp>=0.1.0",
    "runcmd-mcp>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/username/mcp-tools-collection"
Repository = "https://github.com/username/mcp-tools-collection"
Documentation = "https://github.com/username/mcp-tools-collection/blob/main/README.md"

